+++ jackify/frontends/gui/mixins/__init__.py
@@ -1,0 +1,10 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+"""
+GUI Mixins Package
+Reusable mixins for GUI functionality
+"""
+
+from .operation_lock_mixin import OperationLockMixin
+
+__all__ = ['OperationLockMixin']